functions/requirements.txt
firebase_functions~=0.1.0
Pillow
google-cloud-vision
firebase-admin
google-generativeai
